Simplify setup.py by removing regex search for version, moving out some functions to new file tools.py (not setuptools.py because module named this)
